requirements.txt
pyocclient
Pillow
pyyaml
flask